Extract pure service mapping helpers into dedicated module
